Refactor : generate controller
